Test: Move unit tests to test/unit directory
